requirements.txt
streamlit
pandas
numpy
plotly
google-generativeai
python-dotenv
openpyxl
